+++ release_notes.md
@@ -1,13 +1,12 @@
-# Azure Functions CLI 4.3.0
+# Azure Functions CLI 4.4.0
 
 #### Host Version
 
-- Host Version: 4.1042.100
+- Host Version: 4.1043.200
 - In-Proc Host Version: 4.41.100 (4.841.100, 4.641.100)
 
 #### Changes
 
-- <entry>
 - Add Dockerfile for python 3.13 local build environment (#4611)
 - Add support for Python 3.14 (#4668)
 - Implement preview feature to apply host configuration profiles on 'func init' command (#4675)